Extract client script to separate file and bundle with Deno
- Moves client WebSocket script to client-raw.js
- Adds client.ts for raw import support
- Updates client.js to be a bundled module export
- Adjusts deno.json tasks for bundling and serving
- Removes inlined client script from main.ts
