+++ client.ts
@@ -1,0 +1,3 @@
+// This file is used to generate `client.js` until
+// 'raw imports' are supported in jsr
+export { default } from "./client-raw.js" with { type: "text" };
